requirements.txt
PyQt5>=5.8.1
suds-jurko>=0.6
